split serverclasses into separate files
